+++ server/src/fhir/json/definitions/kbv_bundle.rs
@@ -42,9 +42,9 @@ use super::{
 
 pub struct KbvBundleDef;
 
-#[derive(Serialize, Deserialize)]
 #[serde(rename = "Bundle")]
-pub struct KbvBundleCow<'a>(#[serde(with = "KbvBundleDef")] Cow<'a, KbvBundle>);
+#[derive(Clone, Serialize, Deserialize)]
+pub struct KbvBundleCow<'a>(#[serde(with = "KbvBundleDef")] pub Cow<'a, KbvBundle>);
 
 #[serde(rename = "Bundle")]
 #[serde(rename_all = "camelCase")]