+++ server/src/fhir/xml/definitions/misc/binary.rs
@@ -15,8 +15,7 @@
  *
  */
 
-use base64::{decode, encode};
-use serde::{de::Error, Deserialize, Deserializer, Serialize, Serializer};
+use serde::{Deserialize, Serialize};
 
 #[serde(rename = "Binary")]
 #[serde(rename_all = "camelCase")]
@@ -27,18 +26,5 @@ pub struct BinaryDef {
     pub content_type: String,
 
     #[serde(default)]
-    #[serde(serialize_with = "to_base64")]
-    #[serde(deserialize_with = "from_base64")]
-    pub data: Option<Vec<u8>>,
-}
-
-fn to_base64<S: Serializer>(value: &Option<Vec<u8>>, serializer: S) -> Result<S::Ok, S::Error> {
-    value.as_ref().map(encode).serialize(serializer)
-}
-
-fn from_base64<'de, D: Deserializer<'de>>(deserializer: D) -> Result<Option<Vec<u8>>, D::Error> {
-    Option::<String>::deserialize(deserializer)?
-        .map(decode)
-        .transpose()
-        .map_err(|err| D::Error::custom(format!("invalid base64 string: {}", err)))
+    pub data: Option<String>,
 }