+++ server/src/service/misc/cms.rs
@@ -1,0 +1,77 @@
+/*
+ * Copyright (c) 2020 gematik GmbH
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+use openssl::{
+    pkcs7::{Pkcs7, Pkcs7Flags},
+    stack::Stack,
+    x509::{
+        store::{X509Store, X509StoreBuilder},
+        X509,
+    },
+};
+
+use super::super::error::{Error, RequestError};
+
+pub struct Cms {
+    cert: X509,
+    certs: Stack<X509>,
+    store: X509Store,
+}
+
+impl Cms {
+    pub fn new(cert: X509) -> Result<Self, Error> {
+        let mut certs = Stack::new()?;
+        certs.push(cert.clone())?;
+
+        let mut store = X509StoreBuilder::new()?;
+        store.add_cert(cert.clone())?;
+        let store = store.build();
+
+        Ok(Self { cert, certs, store })
+    }
+
+    pub fn verify(&self, data: &str) -> Result<Vec<u8>, RequestError> {
+        let pkcs7 = if data.starts_with("-----BEGIN PKCS7-----") {
+            Pkcs7::from_pem(data.as_bytes())?
+        } else {
+            let data = format!(
+                "-----BEGIN PKCS7-----\n{}\n-----END PKCS7-----",
+                data.trim()
+            );
+
+            Pkcs7::from_pem(data.as_bytes())?
+        };
+
+        let mut ret = Vec::new();
+
+        pkcs7.verify(
+            &self.certs,
+            &self.store,
+            None,
+            Some(&mut ret),
+            Pkcs7Flags::empty(),
+        )?;
+
+        Ok(ret)
+    }
+}
+
+impl Clone for Cms {
+    fn clone(&self) -> Self {
+        Self::new(self.cert.clone()).unwrap()
+    }
+}